@@ -1,28 +1,28 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <Description>The CQRS WebApi flow</Description>
-        <Authors>Konrad Dąbrowski</Authors>
-        <TargetFramework>net5.0</TargetFramework>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <LangVersion>Latest</LangVersion>
-        <AssemblyName>Arcaim.CQRS.WebApi</AssemblyName>
-        <Title>Arcaim.CQRS.WebApi</Title>
-        <PackageId>Arcaim.CQRS.WebApi</PackageId>
-        <Company>Arcaim</Company>
-        <Version>1.3.0</Version>
-    </PropertyGroup>
+  <PropertyGroup>
+    <Description>The CQRS WebApi flow</Description>
+    <Authors>Konrad Dąbrowski</Authors>
+    <TargetFramework>net6.0</TargetFramework>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <LangVersion>Latest</LangVersion>
+    <AssemblyName>Arcaim.CQRS.WebApi</AssemblyName>
+    <Title>Arcaim.CQRS.WebApi</Title>
+    <PackageId>Arcaim.CQRS.WebApi</PackageId>
+    <Company>Arcaim</Company>
+    <Version>6.0.0</Version>
+  </PropertyGroup>
 
-    <ItemGroup>
-        <FrameworkReference Include="Microsoft.AspNetCore.App">
-            <PrivateAssets>all</PrivateAssets>
-        </FrameworkReference>
-    </ItemGroup>
+  <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App">
+      <PrivateAssets>all</PrivateAssets>
+    </FrameworkReference>
+  </ItemGroup>
 
-    <ItemGroup>
-        <PackageReference Include="Arcaim.CQRS.Commands" Version="1.3.0" />
-        <PackageReference Include="Arcaim.CQRS.Queries" Version="1.3.0" />
-        <PackageReference Include="Arcaim.DI.Scanner" Version="1.0.0" />
-    </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Arcaim.CQRS.Commands" Version="6.0.0" />
+    <PackageReference Include="Arcaim.CQRS.Queries" Version="6.0.0" />
+    <PackageReference Include="Arcaim.DI.Scanner" Version="6.0.0" />
+  </ItemGroup>
 
 </Project>
