@@ -1,14 +1,13 @@
 using System;
 
-namespace Arcaim.CQRS.WebApi.Attributes
-{
-    [AttributeUsage(AttributeTargets.Class, AllowMultiple = true)]
-    public class AuthorizeAttribute : Attribute
-    {
-        public string Roles { get; set; }
+namespace Arcaim.CQRS.WebApi.Attributes;
 
-        public AuthorizeAttribute()
-        {
-        }
-    }
+[AttributeUsage(AttributeTargets.Class, AllowMultiple = true)]
+public class AuthorizeAttribute : Attribute
+{
+  public string Roles { get; set; }
+
+  public AuthorizeAttribute()
+  {
+  }
 }
