@@ -1,9 +1,8 @@
 using System;
 
-namespace Arcaim.CQRS.WebApi.Attributes
+namespace Arcaim.CQRS.WebApi.Attributes;
+
+[AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
+public class ValidateAttribute : Attribute
 {
-    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
-    public class ValidateAttribute : Attribute
-    {
-    }
 }
