@@ -1,20 +1,19 @@
 using Microsoft.AspNetCore.Http;
 
-namespace Arcaim.CQRS.WebApi.Exceptions
+namespace Arcaim.CQRS.WebApi.Exceptions;
+
+public class NoPermissionException : WebApiException
 {
-    public class NoPermissionException : WebApiException
-    {
-        public override string Code => "no_permission";
-        public override int StatusCode => StatusCodes.Status403Forbidden;
+  public override string Code => "no_permission";
+  public override int StatusCode => StatusCodes.Status403Forbidden;
 
-        private NoPermissionException(string message) : base(message)
-        {
-        }
+  private NoPermissionException(string message) : base(message)
+  {
+  }
 
-        private static NoPermissionException CreateNoPermissionException(string message)
-            => new NoPermissionException(message);
-        
-        public static NoPermissionException Create()
-            => CreateNoPermissionException("The user is not authorized to this content");
-    }
+  private static NoPermissionException CreateNoPermissionException(string message)
+      => new NoPermissionException(message);
+  
+  public static NoPermissionException Create()
+    => CreateNoPermissionException("The user is not authorized to this content");
 }
