@@ -1,13 +1,12 @@
-namespace Arcaim.CQRS.WebApi.Exceptions
-{
-    public abstract class WebApiException : System.Exception
-    {
-        public abstract string Code { get; }
-        public abstract int StatusCode { get; }
-        public override string Source { get => "Arcaim.CQRS.WebApi"; }
+namespace Arcaim.CQRS.WebApi.Exceptions;
 
-        public WebApiException(string message) : base(message)
-        {
-        }
-    }
+public abstract class WebApiException : System.Exception
+{
+  public abstract string Code { get; }
+  public abstract int StatusCode { get; }
+  public override string Source { get => "Arcaim.CQRS.WebApi"; }
+
+  public WebApiException(string message) : base(message)
+  {
+  }
 }
