@@ -1,9 +1,8 @@
 using System.Threading.Tasks;
 
-namespace Arcaim.CQRS.WebApi.HttpFilter
+namespace Arcaim.CQRS.WebApi.HttpFilter;
+
+public interface IFilterManager
 {
-    public interface IFilterManager
-    {
-        Task InvokeFiltersAsync<T>(T instance);
-    }
+  Task InvokeFiltersAsync<T>(T instance);
 }
