@@ -1,11 +1,10 @@
 using Microsoft.AspNetCore.Routing;
 
-namespace Arcaim.CQRS.WebApi.Interfaces
+namespace Arcaim.CQRS.WebApi.Interfaces;
+
+public interface IWebApi : ICommandApi, IQueryApi
 {
-    public interface IWebApi : ICommandApi, IQueryApi
-    {
-        IEndpointRouteBuilder EndpointRouteBuilder { get; }
-        string Pattern { get; init; }
-        string PatternAction { get; set; }
-    }
+  IEndpointRouteBuilder EndpointRouteBuilder { get; }
+  string Pattern { get; init; }
+  string PatternAction { get; set; }
 }
