@@ -7,93 +7,92 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 
-namespace Arcaim.CQRS.WebApi
+namespace Arcaim.CQRS.WebApi;
+
+internal sealed class WebApi : IWebApi
 {
-    internal sealed class WebApi : IWebApi
+  public IEndpointRouteBuilder EndpointRouteBuilder { get; }
+  public IApplicationBuilder ApplicationBuilder { get; }
+  private string _pattern;
+  public string Pattern
+  {
+    get => string.IsNullOrWhiteSpace(this.PatternAction) ? _pattern : $"{_pattern}/{PatternAction}";
+    init => _pattern = value;
+  }
+
+  public string PatternAction { get; set; }
+
+  internal WebApi(IEndpointRouteBuilder endpointRouteBuilder, string pattern)
+  {
+    EndpointRouteBuilder = endpointRouteBuilder;
+    Pattern = pattern;
+  }
+
+  public IEndpointConventionBuilder Get(RequestDelegate requestDelegate)
+    => EndpointRouteBuilder.MapGet(Pattern, requestDelegate);
+
+  public IEndpointConventionBuilder Get<T, S>()
+    where T : IQuery<S>
+    where S : class
+    => EndpointRouteBuilder.MapGet(Pattern, async ctx =>
     {
-        public IEndpointRouteBuilder EndpointRouteBuilder { get; }
-        public IApplicationBuilder ApplicationBuilder { get; }
-        private string _pattern;
-        public string Pattern
-        {
-            get => string.IsNullOrWhiteSpace(this.PatternAction) ? _pattern : $"{_pattern}/{PatternAction}";
-            init => _pattern = value;
-        }
+      var instance = await ctx.GetModel<T>();
 
-        public string PatternAction { get; set; }
+      await InvokeFilters(instance);
+      var result = await EndpointRouteBuilder
+        .GetService<IQueryDispatcher>()
+        .DispatchAsync(instance);
 
-        internal WebApi(IEndpointRouteBuilder endpointRouteBuilder, string pattern)
-        {
-            EndpointRouteBuilder = endpointRouteBuilder;
-            Pattern = pattern;
-        }
+      ctx.Return(result);
+    });
 
-        public IEndpointConventionBuilder Get(RequestDelegate requestDelegate)
-            => EndpointRouteBuilder.MapGet(Pattern, requestDelegate);
+  public IEndpointConventionBuilder Post(RequestDelegate requestDelegate)
+    => EndpointRouteBuilder.MapPost(Pattern, requestDelegate);
 
-        public IEndpointConventionBuilder Get<T, S>()
-            where T : IQuery<S>
-            where S : class
-            => EndpointRouteBuilder.MapGet(Pattern, async ctx =>
-            {
-                var instance = await ctx.GetModel<T>();
+  public IEndpointConventionBuilder Post<T>() where T : class, ICommand
+    => EndpointRouteBuilder.MapPost(Pattern, async ctx =>
+    {
+      var instance = await ctx.GetModel<T>();
 
-                await InvokeFilters(instance);
-                var result = await EndpointRouteBuilder
-                    .GetService<IQueryDispatcher>()
-                    .DispatchAsync(instance);
+      await InvokeFilters(instance);
+      await EndpointRouteBuilder
+        .GetService<ICommandDispatcher>()
+        .DispatchAsync(instance);
+    });
 
-                ctx.Return(result);
-            });
+  public IEndpointConventionBuilder Put(RequestDelegate requestDelegate)
+    => EndpointRouteBuilder.MapPut(Pattern, requestDelegate);
 
-        public IEndpointConventionBuilder Post(RequestDelegate requestDelegate)
-            => EndpointRouteBuilder.MapPost(Pattern, requestDelegate);
+  public IEndpointConventionBuilder Put<T>() where T : class, ICommand
+    => EndpointRouteBuilder.MapPut(Pattern, async ctx =>
+    {
+      var instance = await ctx.GetModel<T>();
 
-        public IEndpointConventionBuilder Post<T>() where T : class, ICommand
-            => EndpointRouteBuilder.MapPost(Pattern, async ctx =>
-            {
-                var instance = await ctx.GetModel<T>();
+      await InvokeFilters(instance);
+      await EndpointRouteBuilder
+        .GetService<ICommandDispatcher>()
+        .DispatchAsync(instance);
+    });
 
-                await InvokeFilters(instance);
-                await EndpointRouteBuilder
-                    .GetService<ICommandDispatcher>()
-                    .DispatchAsync(instance);
-            });
+  public IEndpointConventionBuilder Delete(RequestDelegate requestDelegate)
+    => EndpointRouteBuilder.MapDelete(Pattern, requestDelegate);
 
-        public IEndpointConventionBuilder Put(RequestDelegate requestDelegate)
-            => EndpointRouteBuilder.MapPut(Pattern, requestDelegate);
+  public IEndpointConventionBuilder Delete<T>() where T : class, ICommand
+    => EndpointRouteBuilder.MapDelete(Pattern, async ctx =>
+    {
+      var instance = await ctx.GetModel<T>();
 
-        public IEndpointConventionBuilder Put<T>() where T : class, ICommand
-            => EndpointRouteBuilder.MapPut(Pattern, async ctx =>
-            {
-                var instance = await ctx.GetModel<T>();
+      await InvokeFilters(instance);
+      await EndpointRouteBuilder.GetService<ICommandDispatcher>()
+        .DispatchAsync(instance);
+    });
 
-                await InvokeFilters(instance);
-                await EndpointRouteBuilder
-                    .GetService<ICommandDispatcher>()
-                    .DispatchAsync(instance);
-            });
-
-        public IEndpointConventionBuilder Delete(RequestDelegate requestDelegate)
-            => EndpointRouteBuilder.MapDelete(Pattern, requestDelegate);
-
-        public IEndpointConventionBuilder Delete<T>() where T : class, ICommand
-            => EndpointRouteBuilder.MapDelete(Pattern, async ctx =>
-            {
-                var instance = await ctx.GetModel<T>();
-
-                await InvokeFilters(instance);
-                await EndpointRouteBuilder.GetService<ICommandDispatcher>()
-                    .DispatchAsync(instance);
-            });
-
-        private async Task InvokeFilters<T>(T instance)
-        {
-            var filterManager = EndpointRouteBuilder.GetService<IFilterManager>();
-            if (filterManager is not null)
-            {
-                await filterManager.InvokeFiltersAsync(instance);
-            }
-        }
+  private async Task InvokeFilters<T>(T instance)
+  {
+    var filterManager = EndpointRouteBuilder.GetService<IFilterManager>();
+    if (filterManager is not null)
+    {
+      await filterManager.InvokeFiltersAsync(instance);
     }
+  }
 }
